refactor(core): single-source pure statusline primitives in _shared (#143)
Extract the duplicated constants and leaf functions behind the parity
suite into src/claude_statusline/_shared.py: colors, MI/zone/pacman
logic, tok/s compute, layout helpers, stdin trust-boundary helpers,
session/CSV validation, lock + rotate cores, and git-info subprocess
core. Package modules import from _shared under their historical names;
behavior is unchanged and the parity suite stays green (425/425).
