+++ src/claude_statusline/core/colors.py
@@ -2,72 +2,11 @@
 
 from __future__ import annotations
 
-import re
-
-# ANSI color codes (defaults)
-BLUE = "\033[0;34m"
-MAGENTA = "\033[0;35m"
-CYAN = "\033[0;36m"
-GREEN = "\033[0;32m"
-YELLOW = "\033[0;33m"
-RED = "\033[0;31m"
-BOLD = "\033[1m"
-DIM = "\033[2m"
-RESET = "\033[0m"
-
-# Mapping from color names to ANSI codes
-COLOR_NAMES: dict[str, str] = {
-    "black": "\033[0;30m",
-    "red": "\033[0;31m",
-    "green": "\033[0;32m",
-    "yellow": "\033[0;33m",
-    "blue": "\033[0;34m",
-    "magenta": "\033[0;35m",
-    "cyan": "\033[0;36m",
-    "white": "\033[0;37m",
-    "bright_black": "\033[0;90m",
-    "bright_red": "\033[0;91m",
-    "bright_green": "\033[0;92m",
-    "bright_yellow": "\033[0;93m",
-    "bright_blue": "\033[0;94m",
-    "bright_magenta": "\033[0;95m",
-    "bright_cyan": "\033[0;96m",
-    "bright_white": "\033[0;97m",
-    "bold_white": "\033[1;97m",
-    "dim": "\033[2m",
-}
-
-_HEX_RE = re.compile(r"^#([0-9a-fA-F]{6})$")
-
-
-def parse_color(value: str) -> str | None:
-    """Parse a color value into an ANSI escape code.
-
-    Accepts:
-      - Named colors: "red", "green", "bright_cyan", etc.
-      - Hex colors: "#ff5733" (converted to 24-bit ANSI)
-
-    Returns:
-        ANSI escape code string, or None if the value is not recognized.
-    """
-    value = value.strip().lower()
-    if not value:
-        return None
-
-    # Named color
-    if value in COLOR_NAMES:
-        return COLOR_NAMES[value]
-
-    # Hex color (#rrggbb)
-    m = _HEX_RE.match(value)
-    if m:
-        hex_str = m.group(1)
-        r = int(hex_str[0:2], 16)
-        g = int(hex_str[2:4], 16)
-        b = int(hex_str[4:6], 16)
-        return f"\033[38;2;{r};{g};{b}m"
-
-    return None
+# Single-sourced in _shared (Task 5.2, F-DEAD-001): the standalone script loads
+# the same bodies from claude_statusline._shared / its vendored copy.
+from claude_statusline._shared import BLUE, BOLD, CYAN, DIM, GREEN, MAGENTA, RED, RESET, YELLOW
+from claude_statusline._shared import COLOR_NAMES as COLOR_NAMES
+from claude_statusline._shared import parse_color as parse_color
 
 
 class ColorManager:
