+++ src/claude_statusline/graphs/statistics.py
@@ -4,6 +4,16 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
+# tok/s + compaction primitives are single-sourced in _shared (Task 5.2,
+# F-DEAD-001); the standalone script loads the same bodies from
+# claude_statusline._shared / its vendored copy.
+from claude_statusline._shared import (
+    COMPACTION_DROP_THRESHOLD as COMPACTION_DROP_THRESHOLD,
+)
+from claude_statusline._shared import compute_tps as compute_tps
+from claude_statusline._shared import detect_compaction_events as detect_compaction_events
+from claude_statusline._shared import format_tps as format_tps
+
 
 @dataclass
 class Stats:
@@ -71,37 +81,10 @@ def detect_spike(deltas: list[int], context_window_size: int, window: int = 5) -
     return False
 
 
-def detect_compaction_events(values: list[int], drop_threshold: float = 0.5) -> list[int]:
-    """Detect compaction events in a list of token counts.
-
-    A compaction event is identified when ``values[i] < values[i-1] * (1 - drop_threshold)``,
-    i.e., the context dropped by more than *drop_threshold* fraction in a single step.
-    With the default threshold of 0.5 this means the new value is less than half the old value.
-
-    The parameter ``drop_threshold`` controls what fraction of context must be lost to count
-    as a compaction.  Increasing it makes detection stricter (only very large drops qualify);
-    decreasing it makes it more sensitive.
-
-    Args:
-        values: Sequence of token counts (e.g., ``current_used_tokens`` over time).
-        drop_threshold: Fraction of context that must be lost to qualify as compaction
-                        (default: 0.5, i.e., > 50 % drop).
-
-    Returns:
-        List of indices *i* (into *values*) where a compaction was detected.
-        Indices are 1-based (the earliest possible index is 1).
-    """
-    if len(values) < 2:
-        return []
-
-    events: list[int] = []
-    for i in range(1, len(values)):
-        prev = values[i - 1]
-        curr = values[i]
-        # Guard against zero-division; if prev == 0 there is nothing to compare
-        if prev > 0 and curr < prev * (1.0 - drop_threshold):
-            events.append(i)
-    return events
+# detect_compaction_events / compute_tps / format_tps bodies are single-sourced
+# in _shared (imported above). The default-threshold semantics are preserved:
+# the shared detect_compaction_events resolves drop_threshold=None to
+# COMPACTION_DROP_THRESHOLD (0.5), matching the former 0.5 literal default.
 
 
 def calculate_deltas(values: list[int]) -> list[int]:
@@ -123,76 +106,6 @@ def calculate_deltas(values: list[int]) -> list[int]:
         deltas.append(max(0, delta))
 
     return deltas
-
-
-def compute_tps(
-    samples: list[tuple[int, int]],
-    window: int = 5,
-) -> float | None:
-    """Compute a smoothed, session-rolling model throughput in tokens/second.
-
-    Rather than the jumpy per-turn instantaneous speed (which swings between,
-    say, 1.5 and 80 tok/s depending on how many tokens a single turn happened
-    to emit), this returns a **rolling, token-weighted average** over the most
-    recent turns. The average is weighted by output tokens, so a tiny 3-token
-    turn cannot drag the number down the way a plain mean-of-ratios would —
-    the result tracks the genuine "speed of the model" across the session.
-
-    Each ``sample`` is an ``(output_tokens, api_duration_ms)`` pair taken from
-    a state row (plus the live reading), where ``api_duration_ms`` is the
-    cumulative ``cost.total_api_duration_ms`` ("time spent waiting for API
-    responses" — it excludes user idle time, tool execution, and thinking).
-    A *turn* is the transition between two consecutive samples: its output is
-    that row's ``current_usage.output_tokens`` and its API time is the delta
-    of the cumulative durations. Turns with a non-positive API-time delta
-    (same response refreshed twice) or non-positive output are dropped.
-
-    The average over the last ``window`` valid turns is token-weighted:
-
-        tok/s = Σ output_tokens / (Σ api_time_ms / 1000)
-
-    Because both sums accumulate over the kept turns, a turn that contributes
-    no valid sample simply isn't in the sums — the previously established
-    average persists ("keep last average" on missing data) as long as at least
-    one valid turn remains in the window.
-
-    Args:
-        samples: Chronological ``(output_tokens, api_duration_ms)`` pairs, one
-            per state row, with the live reading last. ``api_duration_ms`` is
-            the cumulative API wait time at that row.
-        window: Number of most-recent valid turns to average over (>= 1).
-
-    Returns:
-        Rolling throughput in tokens/second, or ``None`` when no valid turn
-        exists yet (first row, all legacy rows, or no real API time elapsed).
-        ``None`` signals "hide the display".
-    """
-    if window < 1:
-        window = 1
-
-    # Reconstruct per-turn (output, api_time_ms) from consecutive samples,
-    # keeping only turns with real elapsed API time and real output.
-    turns: list[tuple[int, int]] = []
-    for (_, prev_dur), (out, cur_dur) in zip(samples, samples[1:], strict=False):
-        # A zero/negative previous cumulative means a legacy row without the
-        # field — differencing against it would understate throughput badly.
-        if prev_dur <= 0:
-            continue
-        delta_ms = cur_dur - prev_dur
-        if delta_ms <= 0 or out <= 0:
-            continue
-        turns.append((out, delta_ms))
-
-    if not turns:
-        return None
-
-    recent = turns[-window:]
-    total_output = sum(out for out, _ in recent)
-    total_ms = sum(ms for _, ms in recent)
-    if total_ms <= 0:
-        return None
-
-    return total_output / (total_ms / 1000.0)
 
 
 def compute_tps_series(
@@ -240,18 +153,3 @@ def compute_tps_series(
             continue
         series.append((i, tps))
     return series
-
-
-def format_tps(tps: float, precision: int = 1, unit: str = "tok/s") -> str:
-    """Format a tokens-per-second value for display.
-
-    Args:
-        tps: Throughput in tokens per second.
-        precision: Number of decimal places (clamped to the range 0..10).
-        unit: Unit label appended after the value (e.g. ``"tok/s"``).
-
-    Returns:
-        Formatted string like ``"42.5 tok/s"``.
-    """
-    precision = min(10, max(0, precision))
-    return f"{tps:.{precision}f} {unit}"
